Remade ticket updates system, all controllers now authorize actions directly in the methods

+++ app/Http/Controllers/LabelController.php
@@ -7,6 +7,7 @@ use App\Http\Requests\UpdateLabelRequest;
 use App\Http\Resources\LabelResource;
 use App\Models\Label;
 use App\Models\Project;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\Request;
 use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
 use Illuminate\Http\Response;
@@ -14,11 +15,6 @@ use Throwable;
 
 class LabelController extends Controller
 {
-    public function __construct()
-    {
-        $this->authorizeResource(Label::class, 'label');
-    }
-
     /**
      * Display a listing of the resource.
      *
@@ -34,12 +30,15 @@ class LabelController extends Controller
      * Store a newly created resource in storage.
      *
      * @param StoreLabelRequest $request
+     * @param Project $project
      * @return Response
+     * @throws AuthorizationException
      */
-    public function store(StoreLabelRequest $request, Project $project)
+    public function store(StoreLabelRequest $request, Project $project): Response
     {
+        $this->authorize('update', $project);
         $label = new Label();
-        $label->project_id = $project->id;
+        $label->project()->associate($project);
         $label->fill($request->all());
         $label->save();
         
@@ -51,9 +50,11 @@ class LabelController extends Controller
      *
      * @param Label $label
      * @return LabelResource|Response
+     * @throws AuthorizationException
      */
     public function show(Label $label): Response|LabelResource
     {
+        $this->authorize('view', $label);
         return new LabelResource($label);
     }
 
@@ -63,9 +64,11 @@ class LabelController extends Controller
      * @param UpdateLabelRequest $request
      * @param Label $label
      * @return Response
+     * @throws AuthorizationException
      */
     public function update(UpdateLabelRequest $request, Label $label): Response
     {
+        $this->authorize('update', $label);
         $label->fill($request->all());
         $label->save();
         
@@ -81,6 +84,7 @@ class LabelController extends Controller
      */
     public function destroy(Label $label): Response
     {
+        $this->authorize('delete', $label);
         $label->deleteOrFail();
         
         return response('OK');
